requirements.txt
crewai==0.86.0
crewai-tools==0.17.0
langchain==0.3.11
langchain-community==0.3.11
langchain-text-splitters==0.3.2
chromadb==0.5.23
mem0ai==0.1.34
groq==0.13.0
